Rename Core components files
