--- packages/ra-core/src/ui/delete/index.js
@@ -1,1 +1,0 @@
-export CoreDelete from './Delete';
