--- packages/ra-core/src/ui/detail/index.js
@@ -1,3 +1,0 @@
-export CoreCreate from './Create';
-export CoreEdit from './Edit';
-export CoreShow from './Show';
