+++ packages/ra-core/src/ui/index.js
@@ -1,6 +1,9 @@
-export * from './detail';
-export * from './delete';
+export CoreCreate from './CoreCreate';
+export CoreDelete from './CoreDelete';
+export CoreEdit from './CoreEdit';
+export CoreList from './CoreList';
+export CoreShow from './CoreShow';
+
 export * from './field';
 export * from './form';
 export * from './input';
-export * from './list';
