--- packages/ra-core/src/ui/list/index.js
@@ -1,1 +1,0 @@
-export CoreList from './List';
